criando a classe caixa e fazendo ela herdar nome e id do funcionario
