+++ include/caixa.h
@@ -1,21 +1,18 @@
-#ifndef FUNCIONARIO_H
-#define FUNCIONARIO_H
+#ifndef CAIXA_H
+#define CAIXA_H
 #include "cliente.h"
 
 #include <string>
 
 using namespace std;
 
-class Funcionario {
+class Caixa : public Funcionario{
     private:
-        string _nome;
-        int _id;
 
     public:
-        Funcionario(string nome, int id);
-        string getNome();
-        void setNome(string nome);
-        string emitirNotaFiscal(Cliente cliente);   
+        Caixa(string nome, int id);
+        string emitirNotaFiscal(Cliente cliente);
+         
 };
