+++ src/caixa.cpp
@@ -1,20 +1,11 @@
 #include "../include/funcionario.h"
 #include <string>
 
-Funcionario::Funcionario(string nome, int id) {
+Caixa::Caixa(string nome, int id) : Funcionario(nome, id) {
 
-    _nome = nome;
-    _id = id;
+
 }
 
-string Funcionario::getNome() {
-    return _nome;
-}
-
-void Funcionario::setNome(string nome) {
-    _nome = nome;
-}
-
-string Funcionario::emitirNotaFiscal(Cliente cliente) {
+string Caixa::emitirNotaFiscal(Cliente cliente) {
     return 0;
 }
