+++ src/main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-    Funcionario f = Funcionario("Joao", 1);
+    Caixa f = Caixa("Joao", 1);
     cout << f.getNome() << endl;
     return 0;
 }
